fix: view transition in presenter mode
